Time: 0 ms (100.00%) | Memory: 8.5 MB (91.92%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -31,7 +31,7 @@ public:
         if(n>0)
         return ans;
         else
-        return (double)1.0/(ans);
+        return 1.0/(ans);
